+++ lib/resourcebuilder/core.go
@@ -26,10 +26,14 @@ func (b *serviceAccountBuilder) WithModifier(f MetaV1ObjectModifierFunc) Interfa
 	return b
 }
 
-func (b *serviceAccountBuilder) Do() error {
+func (b *serviceAccountBuilder) Do(initial bool) error {
 	serviceAccount := resourceread.ReadServiceAccountV1OrDie(b.raw)
 	if b.modifier != nil {
 		b.modifier(serviceAccount)
+	}
+	if initial {
+		_, err := b.client.ServiceAccounts(serviceAccount.Namespace).Create(serviceAccount)
+		return err
 	}
 	_, _, err := resourceapply.ApplyServiceAccount(b.client, serviceAccount)
 	return err
@@ -53,10 +57,14 @@ func (b *configMapBuilder) WithModifier(f MetaV1ObjectModifierFunc) Interface {
 	return b
 }
 
-func (b *configMapBuilder) Do() error {
+func (b *configMapBuilder) Do(initial bool) error {
 	configMap := resourceread.ReadConfigMapV1OrDie(b.raw)
 	if b.modifier != nil {
 		b.modifier(configMap)
+	}
+	if initial {
+		_, err := b.client.ConfigMaps(configMap.Namespace).Create(configMap)
+		return err
 	}
 	_, _, err := resourceapply.ApplyConfigMap(b.client, configMap)
 	return err
@@ -80,10 +88,14 @@ func (b *namespaceBuilder) WithModifier(f MetaV1ObjectModifierFunc) Interface {
 	return b
 }
 
-func (b *namespaceBuilder) Do() error {
+func (b *namespaceBuilder) Do(initial bool) error {
 	namespace := resourceread.ReadNamespaceV1OrDie(b.raw)
 	if b.modifier != nil {
 		b.modifier(namespace)
+	}
+	if initial {
+		_, err := b.client.Namespaces().Create(namespace)
+		return err
 	}
 	_, _, err := resourceapply.ApplyNamespace(b.client, namespace)
 	return err
@@ -107,10 +119,14 @@ func (b *serviceBuilder) WithModifier(f MetaV1ObjectModifierFunc) Interface {
 	return b
 }
 
-func (b *serviceBuilder) Do() error {
+func (b *serviceBuilder) Do(initial bool) error {
 	service := resourceread.ReadServiceV1OrDie(b.raw)
 	if b.modifier != nil {
 		b.modifier(service)
+	}
+	if initial {
+		_, err := b.client.Services(service.Namespace).Create(service)
+		return err
 	}
 	_, _, err := resourceapply.ApplyService(b.client, service)
 	return err
