+++ pkg/cvo/cvo.go
@@ -479,7 +479,7 @@ func (b *resourceBuilder) BuilderFor(m *lib.Manifest) (resourcebuilder.Interface
 	return internal.NewGenericBuilder(client, *m)
 }
 
-func (b *resourceBuilder) Apply(m *lib.Manifest) error {
+func (b *resourceBuilder) Apply(m *lib.Manifest, initial bool) error {
 	builder, err := b.BuilderFor(m)
 	if err != nil {
 		return err
@@ -487,5 +487,5 @@ func (b *resourceBuilder) Apply(m *lib.Manifest) error {
 	if b.modifier != nil {
 		builder = builder.WithModifier(b.modifier)
 	}
-	return builder.Do()
+	return builder.Do(initial)
 }
